Separate the internal and external lifecycle hooks tests
