+++ lifecycle.go
@@ -475,13 +475,13 @@ func (p *DockerProvider) preCreateContainerHook(ctx context.Context, req Contain
 	return nil
 }
 
-// CombineContainerHooks it returns just one ContainerLifecycle hook, as the result of combining
+// combineContainerHooks it returns just one ContainerLifecycle hook, as the result of combining
 // the default hooks with the user-defined hooks. The function will loop over all the default hooks,
 // storing each of the hooks in a slice, and then it will loop over all the user-defined hooks,
 // appending or prepending them to the slice of hooks. The order of hooks is the following:
 // - for Pre-hooks, always run the default hooks first, then append the user-defined hooks
 // - for Post-hooks, always run the user-defined hooks first, then the default hooks
-func CombineContainerHooks(defaultHooks, userDefinedHooks []ContainerLifecycleHooks) ContainerLifecycleHooks {
+func combineContainerHooks(defaultHooks, userDefinedHooks []ContainerLifecycleHooks) ContainerLifecycleHooks {
 	preCreates := []ContainerRequestHook{}
 	postCreates := []ContainerHook{}
 	preStarts := []ContainerHook{}
