+++ lifecycle_hooks_test.go
@@ -1,0 +1,247 @@
+// This test is testing very internal logic that should not be exported away from this package. We'll
+// leave it in the main testcontainers package. Do not use for user facing examples.
+package testcontainers
+
+import (
+	"bufio"
+	"context"
+	"fmt"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/testcontainers/testcontainers-go/wait"
+)
+
+func TestCombineLifecycleHooks(t *testing.T) {
+	prints := []string{}
+
+	preCreateFunc := func(prefix string, hook string, lifecycleID int, hookID int) func(ctx context.Context, req ContainerRequest) error {
+		return func(ctx context.Context, _ ContainerRequest) error {
+			prints = append(prints, fmt.Sprintf("[%s] pre-%s hook %d.%d", prefix, hook, lifecycleID, hookID))
+			return nil
+		}
+	}
+	hookFunc := func(prefix string, hookType string, hook string, lifecycleID int, hookID int) func(ctx context.Context, c Container) error {
+		return func(ctx context.Context, _ Container) error {
+			prints = append(prints, fmt.Sprintf("[%s] %s-%s hook %d.%d", prefix, hookType, hook, lifecycleID, hookID))
+			return nil
+		}
+	}
+	preFunc := func(prefix string, hook string, lifecycleID int, hookID int) func(ctx context.Context, c Container) error {
+		return hookFunc(prefix, "pre", hook, lifecycleID, hookID)
+	}
+	postFunc := func(prefix string, hook string, lifecycleID int, hookID int) func(ctx context.Context, c Container) error {
+		return hookFunc(prefix, "post", hook, lifecycleID, hookID)
+	}
+
+	lifecycleHookFunc := func(prefix string, lifecycleID int) ContainerLifecycleHooks {
+		return ContainerLifecycleHooks{
+			PreCreates:     []ContainerRequestHook{preCreateFunc(prefix, "create", lifecycleID, 1), preCreateFunc(prefix, "create", lifecycleID, 2)},
+			PostCreates:    []ContainerHook{postFunc(prefix, "create", lifecycleID, 1), postFunc(prefix, "create", lifecycleID, 2)},
+			PreStarts:      []ContainerHook{preFunc(prefix, "start", lifecycleID, 1), preFunc(prefix, "start", lifecycleID, 2)},
+			PostStarts:     []ContainerHook{postFunc(prefix, "start", lifecycleID, 1), postFunc(prefix, "start", lifecycleID, 2)},
+			PostReadies:    []ContainerHook{postFunc(prefix, "ready", lifecycleID, 1), postFunc(prefix, "ready", lifecycleID, 2)},
+			PreStops:       []ContainerHook{preFunc(prefix, "stop", lifecycleID, 1), preFunc(prefix, "stop", lifecycleID, 2)},
+			PostStops:      []ContainerHook{postFunc(prefix, "stop", lifecycleID, 1), postFunc(prefix, "stop", lifecycleID, 2)},
+			PreTerminates:  []ContainerHook{preFunc(prefix, "terminate", lifecycleID, 1), preFunc(prefix, "terminate", lifecycleID, 2)},
+			PostTerminates: []ContainerHook{postFunc(prefix, "terminate", lifecycleID, 1), postFunc(prefix, "terminate", lifecycleID, 2)},
+		}
+	}
+
+	defaultHooks := []ContainerLifecycleHooks{lifecycleHookFunc("default", 1), lifecycleHookFunc("default", 2)}
+	userDefinedHooks := []ContainerLifecycleHooks{lifecycleHookFunc("user-defined", 1), lifecycleHookFunc("user-defined", 2), lifecycleHookFunc("user-defined", 3)}
+
+	hooks := combineContainerHooks(defaultHooks, userDefinedHooks)
+
+	// call all the hooks in the right order, honouring the lifecycle
+
+	req := ContainerRequest{}
+	err := hooks.Creating(context.Background())(req)
+	require.NoError(t, err)
+
+	c := &DockerContainer{}
+
+	err = hooks.Created(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Starting(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Started(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Readied(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Stopping(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Stopped(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Terminating(context.Background())(c)
+	require.NoError(t, err)
+	err = hooks.Terminated(context.Background())(c)
+	require.NoError(t, err)
+
+	// assertions
+
+	// There are 2 default container lifecycle hooks and 3 user-defined container lifecycle hooks.
+	// Each lifecycle hook has 2 pre-create hooks and 2 post-create hooks.
+	// That results in 16 hooks per lifecycle (8 defaults + 12 user-defined = 20)
+
+	// There are 5 lifecycles (create, start, ready, stop, terminate),
+	// but ready has only half of the hooks (it only has post), so we have 90 hooks in total.
+	assert.Len(t, prints, 90)
+
+	// The order of the hooks is:
+	// - pre-X hooks: first default (2*2), then user-defined (3*2)
+	// - post-X hooks: first user-defined (3*2), then default (2*2)
+
+	for i := 0; i < 5; i++ {
+		var hookType string
+		// this is the particular order of execution for the hooks
+		switch i {
+		case 0:
+			hookType = "create"
+		case 1:
+			hookType = "start"
+		case 2:
+			hookType = "ready"
+		case 3:
+			hookType = "stop"
+		case 4:
+			hookType = "terminate"
+		}
+
+		initialIndex := i * 20
+		if i >= 2 {
+			initialIndex -= 10
+		}
+
+		if hookType != "ready" {
+			// default pre-hooks: 4 hooks
+			assert.Equal(t, fmt.Sprintf("[default] pre-%s hook 1.1", hookType), prints[initialIndex])
+			assert.Equal(t, fmt.Sprintf("[default] pre-%s hook 1.2", hookType), prints[initialIndex+1])
+			assert.Equal(t, fmt.Sprintf("[default] pre-%s hook 2.1", hookType), prints[initialIndex+2])
+			assert.Equal(t, fmt.Sprintf("[default] pre-%s hook 2.2", hookType), prints[initialIndex+3])
+
+			// user-defined pre-hooks: 6 hooks
+			assert.Equal(t, fmt.Sprintf("[user-defined] pre-%s hook 1.1", hookType), prints[initialIndex+4])
+			assert.Equal(t, fmt.Sprintf("[user-defined] pre-%s hook 1.2", hookType), prints[initialIndex+5])
+			assert.Equal(t, fmt.Sprintf("[user-defined] pre-%s hook 2.1", hookType), prints[initialIndex+6])
+			assert.Equal(t, fmt.Sprintf("[user-defined] pre-%s hook 2.2", hookType), prints[initialIndex+7])
+			assert.Equal(t, fmt.Sprintf("[user-defined] pre-%s hook 3.1", hookType), prints[initialIndex+8])
+			assert.Equal(t, fmt.Sprintf("[user-defined] pre-%s hook 3.2", hookType), prints[initialIndex+9])
+		}
+
+		// user-defined post-hooks: 6 hooks
+		assert.Equal(t, fmt.Sprintf("[user-defined] post-%s hook 1.1", hookType), prints[initialIndex+10])
+		assert.Equal(t, fmt.Sprintf("[user-defined] post-%s hook 1.2", hookType), prints[initialIndex+11])
+		assert.Equal(t, fmt.Sprintf("[user-defined] post-%s hook 2.1", hookType), prints[initialIndex+12])
+		assert.Equal(t, fmt.Sprintf("[user-defined] post-%s hook 2.2", hookType), prints[initialIndex+13])
+		assert.Equal(t, fmt.Sprintf("[user-defined] post-%s hook 3.1", hookType), prints[initialIndex+14])
+		assert.Equal(t, fmt.Sprintf("[user-defined] post-%s hook 3.2", hookType), prints[initialIndex+15])
+
+		// default post-hooks: 4 hooks
+		assert.Equal(t, fmt.Sprintf("[default] post-%s hook 1.1", hookType), prints[initialIndex+16])
+		assert.Equal(t, fmt.Sprintf("[default] post-%s hook 1.2", hookType), prints[initialIndex+17])
+		assert.Equal(t, fmt.Sprintf("[default] post-%s hook 2.1", hookType), prints[initialIndex+18])
+		assert.Equal(t, fmt.Sprintf("[default] post-%s hook 2.2", hookType), prints[initialIndex+19])
+	}
+}
+
+func TestLifecycleHooks_WithMultipleHooks(t *testing.T) {
+	ctx := context.Background()
+
+	dl := linesTestLogger{}
+
+	req := ContainerRequest{
+		Image: nginxAlpineImage,
+		LifecycleHooks: []ContainerLifecycleHooks{
+			DefaultLoggingHook(&dl),
+			DefaultLoggingHook(&dl),
+		},
+	}
+
+	c, err := GenericContainer(ctx, GenericContainerRequest{
+		ContainerRequest: req,
+		Started:          true,
+	})
+	require.NoError(t, err)
+	require.NotNil(t, c)
+
+	duration := 1 * time.Second
+	err = c.Stop(ctx, &duration)
+	require.NoError(t, err)
+
+	err = c.Start(ctx)
+	require.NoError(t, err)
+
+	err = c.Terminate(ctx)
+	require.NoError(t, err)
+
+	require.Len(t, dl.data, 24)
+}
+
+type linesTestLogger struct {
+	data []string
+}
+
+func (l *linesTestLogger) Printf(format string, args ...interface{}) {
+	l.data = append(l.data, fmt.Sprintf(format, args...))
+}
+
+func TestPrintContainerLogsOnError(t *testing.T) {
+	ctx := context.Background()
+
+	req := ContainerRequest{
+		Image:      "docker.io/alpine",
+		Cmd:        []string{"echo", "-n", "I am expecting this"},
+		WaitingFor: wait.ForLog("I was expecting that").WithStartupTimeout(5 * time.Second),
+	}
+
+	arrayOfLinesLogger := linesTestLogger{
+		data: []string{},
+	}
+
+	container, err := GenericContainer(ctx, GenericContainerRequest{
+		ProviderType:     providerType,
+		ContainerRequest: req,
+		Logger:           &arrayOfLinesLogger,
+		Started:          true,
+	})
+	// it should fail because the waiting for condition is not met
+	if err == nil {
+		t.Fatal(err)
+	}
+	terminateContainerOnEnd(t, ctx, container)
+
+	containerLogs, err := container.Logs(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer containerLogs.Close()
+
+	// read container logs line by line, checking that each line is present in the stdout
+	rd := bufio.NewReader(containerLogs)
+	for {
+		line, err := rd.ReadString('\n')
+		if err != nil {
+			if err.Error() == "EOF" {
+				break
+			}
+
+			t.Fatal("Read Error:", err)
+		}
+
+		// the last line of the array should contain the line of interest,
+		// but we are checking all the lines to make sure that is present
+		found := false
+		for _, l := range arrayOfLinesLogger.data {
+			if strings.Contains(l, line) {
+				found = true
+				break
+			}
+		}
+		assert.True(t, found, "container log line not found in the output of the logger: %s", line)
+	}
+}
